refactor!: remove model/axis.py shim, canonicalize dt_model.axes, export TIME_AXIS
