--- civic_digital_twins/dt_model/model/axis.py
@@ -1,7 +1,0 @@
-"""Axis identity: role constants and the Axis class."""
-
-# SPDX-License-Identifier: Apache-2.0
-
-from ..axes import DOMAIN, ENSEMBLE, PARAMETER, Axis, AxisRole
-
-__all__ = ["AxisRole", "DOMAIN", "PARAMETER", "ENSEMBLE", "Axis"]
